requirements.txt
boto3>=1.34.0
langchain>=0.1.0
langchain-aws>=0.1.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0
aiohttp>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Database and Vector Search
psycopg2-binary>=2.9.7
pgvector>=0.2.4
sqlalchemy>=2.0.0
alembic>=1.12.0

# AI and Embeddings
openai>=1.3.0
sentence-transformers>=2.2.2
tiktoken>=0.5.0
scikit-learn>=1.3.0

